+++ axelor-production/src/main/java/com/axelor/apps/production/exceptions/IExceptionMessage.java
@@ -133,4 +133,9 @@ public interface IExceptionMessage {
 
   static final String MANUF_ORDER_EMAIL_NOT_SENT = /*$$(*/
       "Automatic email was not sent because no default email account was found : please create one." /*)*/;
+
+  static final String UNIT_COST_CALCULATION_IMPORT_FAIL_ERROR = /*$$(*/ "Data import failed" /*)*/;
+
+  static final String UNIT_COST_CALCULATION_IMPORT_CSV_ERROR = /*$$(*/
+      "Uploaded file is not a CSV file" /*)*/;
 }